remove old storage crate
