+++ storage/Cargo.toml
@@ -1,15 +1,20 @@
 [package]
 name = "storage"
-version = "0.1.0"
+version = "0.2.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
-num_cpus = "1.13.0"
-rocksdb = { git = "https://github.com/yiyuanliu/rust-rocksdb", branch = "transaction" }
-tempdir = "0.3.7"
-thiserror = "1.0.30"
+lazy_static = { version = "1.4.0", optional = true }
+num_cpus = { version = "1.13.1", optional = true }
+
+[dependencies.rocksdb]
+git = "https://github.com/yiyuanliu/rust-rocksdb"
+branch = "transaction"
+optional = true
+
+[features]
+default = ["rocksdb_storage"]
+rocksdb_storage = ["rocksdb", "num_cpus", "lazy_static"]
 
 [dev-dependencies]
-tempdir = "0.3.7"
+tempfile = "3.3.0"
