+++ storage/src/rocksdb_storage.rs
@@ -1,0 +1,12 @@
+//! GroveDB storage layer implemented over RocksDB backend.
+mod storage;
+mod storage_context;
+#[cfg(test)]
+mod tests;
+
+pub use storage_context::{
+    PrefixedRocksDbBatch, PrefixedRocksDbRawIterator, PrefixedRocksDbStorageContext,
+    PrefixedRocksDbTransactionContext,
+};
+
+pub use self::storage::RocksDbStorage;
