+++ storage/src/rocksdb_storage/storage_context.rs
@@ -4,10 +4,10 @@ mod context_no_tx;
 mod context_tx;
 mod raw_iterator;
 
-use batch::PrefixedRocksDbBatch;
+pub use batch::PrefixedRocksDbBatch;
 pub use context_no_tx::PrefixedRocksDbStorageContext;
 pub use context_tx::PrefixedRocksDbTransactionContext;
-use raw_iterator::PrefixedRocksDbRawIterator;
+pub use raw_iterator::PrefixedRocksDbRawIterator;
 use rocksdb::{OptimisticTransactionDB, Transaction};
 
 /// Type alias for a database
