--- storage2/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "storage2"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-lazy_static = { version = "1.4.0", optional = true }
-num_cpus = { version = "1.13.1", optional = true }
-
-[dependencies.rocksdb]
-git = "https://github.com/yiyuanliu/rust-rocksdb"
-branch = "transaction"
-optional = true
-
-[features]
-default = ["rocksdb_storage"]
-rocksdb_storage = ["rocksdb", "num_cpus", "lazy_static"]
-
-[dev-dependencies]
-tempfile = "3.3.0"
