--- storage2/src/lib.rs
@@ -1,5 +1,0 @@
-#[cfg(feature = "rocksdb_storage")]
-pub mod rocksdb_storage;
-mod storage;
-
-pub use storage::{Batch, RawIterator, Storage, StorageContext};
