--- storage2/src/rocksdb_storage.rs
@@ -1,8 +1,0 @@
-//! GroveDB storage layer implemented over RocksDB backend.
-mod storage;
-mod storage_context;
-#[cfg(test)]
-mod tests;
-
-pub use storage::RocksDbStorage;
-pub use storage_context::{PrefixedRocksDbStorageContext, PrefixedRocksDbTransactionContext};
